media: cedrus: Add start and stop decoder operations
Some codec needs to perform some additional task when a decoding is started
and stopped, and not only at every frame.

For example, the H264 decoding support needs to allocate buffers that will
be used in the decoding process, but do not need to change over time, or at
each frame.

In order to allow that for codecs, introduce a start and stop hook that
will be called if present at start_streaming and stop_streaming time.

Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>